requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
scikit-learn
pydantic
python-dotenv
supabase
graphviz
python-multipart
